Crate mdbook_epub

A mdbook backend for generating a book in the EPUB format.

Structs

Config
The configuration struct used to tweak how an EPUB document is generated.

Generator
The actual EPUB book renderer.

Constants

DEFAULT_CSS
The default stylesheet used to make the rendered document pretty.

MDBOOK_VERSION
The exact version of mdbook this crate is compiled against.

Functions

generate
Generate an EPUB version of the provided book.

output_filename
Calculate the output filename using the mdbook config.